Update Highlight
The error "Parsing error: Unterminated string constant" suggests there might be an issue with how the string is being defined or terminated. Ensure that the backticks are properly placed and there are no syntax errors elsewhere in the code. Here's the corrected function again:

```javascript
const highlightBackground = (sample) =>  
  `<span style="background-color:yellow;">${sample}</span>`;
```

If this still results in an error, ensure there are no issues with the surrounding code. Here’s a complete example that you can test in isolation:

```javascript
const highlightBackground = (sample) =>  
  `<span style="background-color:yellow;">${sample}</span>`;

console.log(highlightBackground("Test string"));
```

If the error persists, please check for other issues in the surrounding code or ensure your environment is correctly set up to parse ES6 syntax.